requirements.txt
azure-core==1.30.0
azure-functions==1.18.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
idna==3.6
isodate==0.6.1
msal==1.26.0
msal-extensions==1.1.0
packaging==23.2
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.9.0
urllib3==2.2.0
